pyproject.toml
[tool.poetry]
name = "analyzefrc"
version = "0.1.0"
description = ""
authors = ["Tip ten Brink <T.M.tenBrink@student.tudelft.nl>"]

[tool.poetry.dependencies]
python = "~3.9"
frc = "^0.1.0"
matplotlib = "^3.4"
Pillow = "^8.3.2"
loess = "^2.1.1"
deco = "^0.6.3"
readlif = { git = "https://github.com/tmtenbrink/readlif.git" }
pandas = "^1.3.4"
openpyxl = "^3.0.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
